+++ Training/distill.py
@@ -5,7 +5,6 @@ from lm_dataformat import *
 import torch
 import torch.nn.functional as F
 from torch.nn.functional import normalize, cross_entropy
-from torch.nn import DataParallel
 from auto_tqdm import tqdm
 from get_args import get_args
 import deepspeed
@@ -13,6 +12,10 @@ import wandb
 import os
 from random import randint
 import math
+
+#Helper files
+from utility import *
+from loss import *
 
 
 # set random
@@ -44,59 +47,25 @@ model.resize_token_embeddings(len(tokenizer)+2)
 clip_hidden = 512
 #Set up deep speed
 
-#Set up our model wrapper, makes saving easier
-class projection_model(torch.nn.Module):
-    def __init__(self):
-        super(projection_model, self).__init__()
-        self.fc1 = torch.nn.Linear(neo_hidden, neo_hidden//2)
-        self.act = torch.nn.GELU()
-        self.fc2 = torch.nn.Linear(neo_hidden//2, clip_hidden)
-    def forward(self, input_tensor):
-        out = self.act(self.fc1(input_tensor))
-        return self.fc2(out)
+
 
 projection=projection_model()
 projection.half()
 
-class ModelWrapper(torch.nn.Module):
-    def __init__(self, lm, proj):
-        super(ModelWrapper, self).__init__()
-        self.lm = lm
-        self.proj = proj
-    def forward(self, **kwargs):
-        return self.lm(**kwargs)
+
+
 wrapper = ModelWrapper(model, projection)
 model_engine, optimizer, _, _ = deepspeed.initialize(args=args,
                                                         model=wrapper,
                                                         model_parameters=wrapper.parameters())
 model_engine.to(model_engine.local_rank)
-#projection = projection.to(model_engine.local_rank)
+temp_tensor = torch.tensor(temperature).to(model_engine.local_rank)
 
 
 
 #Set up wandb
 if model_engine.local_rank == 0:
     wandb.init(project='speedrun', entity='eleutherai')
-
-
-
-#hparams
-temperature = 1.0
-learning_rate = 5e-5
-weight_decay = 0
-grad_accum = 2
-clip_bs = 20
-pile_bs = 1
-lambda_coeff = 0.1 #relative scale for contrastive loss
-
-mixing_ratio = 3 #Ratio of pile examples to CLIP examples 
-
-# lambda scheduling
-lschedule = "truncated_sine" # truncated_sine, shifted_sine, constant
-lperiod = 1000
-
-temp_tensor = torch.tensor(temperature).to(model_engine.local_rank)
-
 
 if model_engine.local_rank == 0:
     config=wandb.config
@@ -114,133 +83,10 @@ if model_engine.local_rank == 0:
         pass
 
 
-#pytorch dataset for clip juicing
-class DistillDataset(IterableDataset):
-    def __init__(self,\
-        tokenizer, clip_batch_size,
-        clip_dataset_dir, pile_dataset_dir,
-        special_token = "<|CLIP|>", steps = 1e6):
-        self.clip_dataset_dir = clip_dataset_dir
-        self.pile_dataset_dir = pile_dataset_dir
-
-        self.clip_rdr = Reader(self.clip_dataset_dir).stream_data(get_meta=True)
-        self.pile_rdr = Reader(self.pile_dataset_dir).stream_data(get_meta=True)
-
-        #Steps is the total number of elements we should use. Half from CLIP, half from AR
-        self.steps = steps
-        #How many elements are in a single contrastive clip batch
-        self.clip_batch_size = clip_batch_size
-
-        #Start on an example of WIT.
-        self.mix_step = 0 
-
-        #Store special token, add to tokenizer. Remember to resize token embeddings on model!
-        self.tokenizer = tokenizer
-        self.special_token=special_token
-        #Get the index for the special token so that we can adjust the decode mask accordingly.
-        self.special_token_idx=len(self.tokenizer)
-        self.tokenizer.add_tokens([special_token])
-        self.tokenizer.add_special_tokens({'pad_token': '[PAD]'})
-    def __len__(self):
-        return int(self.steps)
-    def __iter__(self):
-        return self
-    def __next__(self):
-        tok = self.tokenizer
-        txts = list()
-        img_latents = list()
-
-        use_clip = False
-        #Mixing
-        self.mix_step += 1
-        if self.mix_step % mixing_ratio==0:
-            use_clip=True
-            self.mix_step=0
-        #Return an element from the pile
-        if not use_clip:
-            for _ in range(pile_bs):
-                text, _ =next(self.pile_rdr)
-        
-                #text split. Check if we are over length. Truncate accordingly
-                ts = text.split()
-                if len(ts) > 1024:
-                    start = randint(0, len(ts)-1024)
-
-                    text = " ".join(ts[start:])
-                txts.append(text)
-
-            #Tokenize text
-            toks = tok.batch_encode_plus(txts, max_length=1024, truncation=True,\
-                padding="max_length", return_tensors="pt").to(model_engine.local_rank)
-            img_latents.append([[0]*clip_hidden])
-        #Return an element from CLIP
-        else:
-            txts = list()
-            for _ in range(self.clip_batch_size):
-                text, img_latent=next(self.clip_rdr)
-                #Append special token
-                text += "<|CLIP|>"
-                txts.append(text)
-                img_latents.append([img_latent])
-
-            #Tokenize text
-            toks = tok.batch_encode_plus(txts, max_length=128, truncation=True,\
-                padding="max_length", return_tensors="pt").to(model_engine.local_rank)
-        #Get the index of the clip tokens.
-        clip_idx = (torch.sum(toks.attention_mask, dim=-1).to("cpu") - torch.tensor([1] * len(txts)))
-        #Get latent vectors
-        latents = torch.cat([torch.tensor(x) for x in img_latents], dim=0).to(model_engine.local_rank)
-        
-        return {
-            **toks,
-            'latent_vecs' : latents,
-            'clip_idx' : clip_idx, 
-            'use_distill' : use_clip,
-        }
-
-#Contrastive loss helper function
-def clip_loss(a, b, temp):
-    # a ~ (b x d)
-    # b ~ (b x d)
-    batch_size, dimension = a.shape
-
-    a_normd = normalize(a, p=2, dim=1).squeeze().to(torch.float32)
-    b_normd = normalize(b, p=2, dim=1).squeeze().to(torch.float32)
-    logits = torch.einsum('i d, j d -> i j', a_normd, b_normd) * temp.exp()
-
-    labels = torch.arange(batch_size).to(model_engine.local_rank)
-
-    loss = cross_entropy(logits, labels) + cross_entropy(logits.T, labels)
-    
-    return loss / 2.0
-
-def ar_loss(out_embeds, inp):
-    # inp :: [b, seq]
-    raw_logits = out_embeds['logits'].squeeze(0)
-    logprobs = F.log_softmax(
-        torch.cat([
-            raw_logits[:, :, :50257],
-            -1e10 * torch.ones(raw_logits.shape[0], raw_logits.shape[1], 2).to(model_engine.local_rank)
-        ], dim=-1)
-        , dim=-1).to(torch.float32)
-    # logprobs :: [b, seq, vocab]
-
-    pred = logprobs[:, :-1]
-    tgt = inp.squeeze(0)[:, 1:]
-
-    is_clip_or_padding_token = tgt >= 50257
-    
-    logits = torch.gather(pred, 2, tgt.unsqueeze(-1)).squeeze(-1) # [batch, seq-1]
-
-    # remove loss of clip-token
-    logits *= 1 - is_clip_or_padding_token.to(torch.int)
-
-    return -logits.sum()
-
 #Load dataset
 data = DistillDataset(tokenizer = tokenizer, clip_batch_size = clip_bs,\
     clip_dataset_dir = "../../clip/",\
-    pile_dataset_dir = "../../pile/")
+    pile_dataset_dir = "../../pile/", local_rank = model_engine.local_rank)
 loader = DataLoader(dataset=data, batch_size=1)
 
 #Check to see if a checkpoint exists. if it does, load that. Otherwise assume we are on step zero.
@@ -250,6 +96,7 @@ try:
     loader_to_step(loader, step+1)
 except:
     step = 0
+
 #Set up progress bar
 pbar = tqdm(enumerate(loader), total=len(data))
 loss_progress = 0.0
@@ -272,15 +119,14 @@ generate_template = tokenizer("EleutherAI is", return_tensors="pt").to(model_eng
 clip_template = tokenizer("A drawing of a goose holding a knife<|CLIP|>", return_tensors="pt").to(model_engine.local_rank)
 clip_idx_template = torch.sum(clip_template['attention_mask'], dim=-1).to("cpu").squeeze().item() - 1
 text_so_far = list()
-#Generate text samples occasionally
+
+#Generate samples occasionally
 def generate_from_template():
     if model_engine.local_rank == 0:
         out_embeds = model_engine(**clip_template, return_dict=True, output_hidden_states=True)['hidden_states'][-4].squeeze()
         CLIP_state = wrapper.proj(out_embeds[clip_idx_template]).cpu().detach().numpy()
         np.savetxt('clip.out', CLIP_state, delimiter=",")
 
-        #text_so_far.append(["EleutherAI is ", text])
-        #wandb.log({"Generated": text})
 
 for batch, data_elem in pbar:
     torch.cuda.empty_cache()
@@ -355,15 +201,11 @@ for batch, data_elem in pbar:
     if (batch+1)%save_every==0:
         torch.distributed.barrier()
         ckpt_id = (batch+1)
-        #model_engine.save_checkpoint(f"models/{wandb.run.name}{ckpt_id}")
-        #model_engine.module.save_pretrained(wandb.run.dir)
+
         # local or global?
         if model_engine.local_rank == 0:
             model_engine.module.lm.save_pretrained(f"models/{wandb.run.name}/{ckpt_id}", ckpt_id)
             torch.save(wrapper.proj, f"models/{wandb.run.name}/projection.pt")
-        #model.save_pretrained("GPT-Neo-Enriched"+str(batch+1))
-        #tokenizer.save_pretrained("GPT-Neo-Enriched"+str(batch+1))
 
 model_engine.save_checkpoint(args.save_dir, ckpt_id, client_sd=client_sd)
-#model.save_pretrained("GPT-Neo-Enriched")
 tokenizer.save_pretrained("GPT-Neo-Enriched")
